Conteudo10

+++ src/Conteudo9/Ex03.java
@@ -6,14 +6,14 @@ public class Ex03 {
 	public static void main(String[] args) {
 		Scanner entrada = new Scanner(System.in);
 		System.out.print("Insira a quantidade de termos: ");
-		int numeroDeTermos = entrada.nextInt();	
-		
+		int numeroDeTermos = entrada.nextInt();
+
 		double somatorio = 0;
-		
+
 		for (int i = 1; i <= numeroDeTermos; i++) {
-			somatorio += ((double)1 + 2*Math.sqrt(i))/(3*i);
+			somatorio += (1 + 2 * Math.sqrt(i)) / (3 * i);
 		}
-		
-		System.out.println("Valor da série = "+somatorio);
+
+		System.out.println("Valor da série = " + somatorio);
 	}
 }
